CLN: Renamed header to separate it from tab related files
